+++ EssentialFeedTests/Feed API/LoadFeedFromRemoteUseCaseTests.swift
@@ -36,7 +36,7 @@ class LoadFeedFromRemoteUseCaseTests: XCTestCase {
         let (sut, client) = makeSUT()
         expect(sut, toCompleteWith: failure(.connectivity)) {
             let clientError = NSError(domain: "Test", code: 0)
-            client.complete(with: clientError)
+            client.completion(with: clientError)
         }
     }
 
@@ -161,36 +161,5 @@ class LoadFeedFromRemoteUseCaseTests: XCTestCase {
         action()
 
         wait(for: [exp], timeout: 3.0)
-    }
-
-    private class HTTPClientSpy: HTTPClient {
-
-        private struct Task: HTTPClientTask {
-            func cancel() { }
-        }
-
-        var requestedURLs: [URL] {
-            messages.map { $0.url }
-        }
-
-        private var messages = [(url: URL, completion: (HTTPClient.Result) -> Void)]()
-
-        func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
-            messages.append((url, completion))
-            return Task()
-        }
-
-        func complete(with error: Error, at index: Int = 0) {
-            messages[index].completion(.failure(error))
-        }
-
-        func complete(withStatus code: Int, data: Data, at index: Int = 0) {
-            let response = HTTPURLResponse(
-                url: requestedURLs[index],
-                statusCode: code,
-                httpVersion: nil,
-                headerFields: nil)!
-            messages[index].completion(.success((data, response)))
-        }
     }
 }
